internal/e2e: Move endpoint tests to featuretests package
Moves the EDS tests from the e2e package to the featuretests package.

Updates #2549
