+++ internal/e2e/e2e.go
@@ -42,7 +42,6 @@ import (
 )
 
 const (
-	endpointType = resource.EndpointType
 	routeType    = resource.RouteType
 	listenerType = resource.ListenerType
 	secretType   = resource.SecretType
